Only restart coordinators if the deployment mode has coordinators.

+++ pkg/deployment/resources/pod_inspector.go
@@ -237,7 +237,7 @@ func (r *Resources) InspectPods(ctx context.Context) (util.Interval, error) {
 	allMembersReady := status.Members.AllMembersReady(spec.GetMode(), spec.Sync.IsEnabled())
 	status.Conditions.Update(api.ConditionTypeReady, allMembersReady, "", "")
 
-	if status.Members.Coordinators.AllFailed() {
+	if spec.GetMode().HasCoordinators() && status.Members.Coordinators.AllFailed() {
 		log.Error().Msg("All coordinators failed - reset")
 		for _, m := range status.Members.Coordinators {
 			if err := r.context.DeletePod(m.PodName); err != nil {
